Changed SPI device setup to return the file descriptor, or -1 for error.
This change aligns SPI device handling with the I2C device handling.
SPI read/write functions need the SPI channel number, which used to be stored as the node->fd
Now the node->fd points to the actual fd, and node->data0 is used to store the SPI channel number.
Curiously, both mcp23s17 and mcp23s08 were already using node->data0 for the SPI channel, so this paradigm was already there.